+++ apps/pie-storybook/stories/pie-cookie-banner.stories.ts
@@ -5,6 +5,7 @@ import { PieLink } from '@justeattakeaway/pie-link';
 import { PieModal } from '@justeattakeaway/pie-modal';
 import { PieIconButton } from '@justeattakeaway/pie-icon-button';
 import { PieToggleSwitch } from '@justeattakeaway/pie-toggle-switch';
+import { action } from '@storybook/addon-actions';
 
 import { type StoryMeta } from '../types';
 import { createStory } from '../utilities';
@@ -23,14 +24,17 @@ const keptReferences = [
 type CookieBannerStoryMeta = StoryMeta<CookieBannerProps>;
 
 const defaultArgs: CookieBannerProps = {
-
+    hasPrimaryActionsOnly: false,
 };
 
 const cookieBannerStoryMeta: CookieBannerStoryMeta = {
     title: 'Cookie Banner',
     component: 'pie-cookie-banner',
     argTypes: {
-
+        hasPrimaryActionsOnly: {
+            description: 'When true, sets the variant to "primary" for the button which accepts necessary cookies only.',
+            control: 'boolean',
+        },
     },
     args: defaultArgs,
     parameters: {
@@ -43,8 +47,21 @@ const cookieBannerStoryMeta: CookieBannerStoryMeta = {
 
 export default cookieBannerStoryMeta;
 
-const BaseStoryTemplate = () : TemplateResult => html`
-        <pie-cookie-banner></pie-cookie-banner>`;
+const necessaryOnlyAction = action('necessary-only');
+const acceptAllAction = action('accept-all');
+const managePrefsAction = action('manage-prefs');
+const prefsSavedAction = action('prefs-saved');
+
+const BaseStoryTemplate = (props: CookieBannerProps) : TemplateResult => {
+    const { hasPrimaryActionsOnly } = props;
+    return html`
+        <pie-cookie-banner
+            ?hasPrimaryActionsOnly="${hasPrimaryActionsOnly}"
+            @pie-cookie-banner-necessary-only="${necessaryOnlyAction}"
+            @pie-cookie-banner-accept-all="${acceptAllAction}"
+            @pie-cookie-banner-manage-prefs="${managePrefsAction}"
+            @pie-cookie-banner-prefs-saved="${prefsSavedAction}"></pie-cookie-banner>`;
+};
 
 /**
  * Creates a 'page' of scrollable HTML. Useful for testing scroll behaviours in a Story.
@@ -61,15 +78,9 @@ const createScrollablePageHTML = () : TemplateResult => {
         <ul>${items}</ul>`;
 };
 
-// TODO: remove the eslint-disable rule when props are added
-// eslint-disable-next-line no-empty-pattern
-const Template = ({}: CookieBannerProps): TemplateResult => html`
-    <pie-cookie-banner></pie-cookie-banner>
-`;
-
-const ScrollablePageStoryTemplate = () : TemplateResult => html`
-    ${BaseStoryTemplate()}
+const ScrollablePageStoryTemplate = (props: CookieBannerProps) : TemplateResult => html`
+    ${BaseStoryTemplate(props)}
     ${createScrollablePageHTML()}`;
 
-export const Default = createStory<CookieBannerProps>(Template, defaultArgs)();
+export const Default = createStory<CookieBannerProps>(BaseStoryTemplate, defaultArgs)();
 export const ScrollablePage = createStory<CookieBannerProps>(ScrollablePageStoryTemplate, defaultArgs)();