+++ packages/components/pie-button/src/index.ts
@@ -2,7 +2,7 @@ import {
     LitElement, html, unsafeCSS, nothing,
 } from 'lit';
 import { property } from 'lit/decorators.js';
-import { validPropertyValues } from '@justeattakeaway/pie-webc-core';
+import { validPropertyValues, defineCustomElement } from '@justeattakeaway/pie-webc-core';
 import {
     ButtonProps, sizes, types, variants, iconPlacements,
 } from './defs';
@@ -15,6 +15,7 @@ export * from './defs';
 const componentSelector = 'pie-button';
 
 /**
+ * @tagname pie-button
  * @slot icon - The icon slot
  * @slot - Default slot
  */
@@ -182,7 +183,7 @@ export class PieButton extends LitElement implements ButtonProps {
     static styles = unsafeCSS(styles);
 }
 
-customElements.define(componentSelector, PieButton);
+defineCustomElement(componentSelector, PieButton);
 
 declare global {
     interface HTMLElementTagNameMap {